feat: implement login/signup forms with improved error handling and UI feedback
