--- next_frontend/src/app/(with-sidebar)/searchengine/page.tsx
@@ -1,23 +1,0 @@
-// import ChatThreadView from "@/components/chats/ChatThreadView";
-
-// const SearchEngine = () => {
-//   return <ChatThreadView threadId="" />;
-// };
-
-// export default SearchEngine;
-"use client";
-
-import { useEffect } from "react";
-import { useRouter } from "next/navigation";
-
-const SearchEngine = () => {
-  const router = useRouter();
-
-  useEffect(() => {
-    router.replace("/chat/new");
-  }, [router]);
-
-  return null;
-};
-
-export default SearchEngine;
